requirements.txt
networkx>=3.1
sentence-transformers>=2.2.0
scikit-learn>=1.0.2
pyyaml>=6.0.0
matplotlib>=3.7.0
torch>=2.0.0
numpy>=1.21.0
click>=8.0.0
pandas>=2.0.0
aiohttp>=3.8.0